Move the graphs protocols to their own file to declutter the GraphCollectionView file.
